requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
pymongo
requests>=2.28.0
linkup
motor
httpx>=0.24.0
mcp
